Add: README UPDATE:   - Change package name   - Replace "Location" into "Rent"   - Replace "Interface" into "Service"
