+++ rmi-bike/src/main/java/interfaces/rent/RentService.java
@@ -1,0 +1,9 @@
+package interfaces.rent;
+
+import java.rmi.Remote;
+import java.rmi.RemoteException;
+import java.util.Date;
+
+public interface RentService extends Remote {
+    Date getEnd() throws RemoteException;
+}
