--- rmi-bike/src/main/java/web/service/rmi/bike/interfaces/location/LocationInterface.java
@@ -1,9 +1,0 @@
-package web.service.rmi.bike.interfaces.location;
-
-import java.rmi.Remote;
-import java.rmi.RemoteException;
-import java.util.Date;
-
-public interface LocationInterface extends Remote {
-    Date getEnd() throws RemoteException;
-}
