+++ rmi-customer/README.md
@@ -11,7 +11,41 @@ All services is binded on port `1099`.
 
 ## `/CustomerListService`
 
-- Retrieve the customer with `uuid`
+### CustomerList
+
+- Get all customers with their uuid. The return value will be a `HashMap<UUID, CustomerService>`.
+```java
+    getAll()
+```
+
+- Get the customer with `uuid`. The return value will be a `Optional<CustomerService>`.
 ```java
     getCustomerByUUID(String uuid)
+```
+
+- Adds a customer to the list with his `firstName`, `lastName`, `customerType`. The return value will be a `void`.
+```java
+    add(String firstName, String lastName, CustomerType customerType)
+```
+
+### Customer
+
+- Get customer `firstName`. The return value will be a `String`.
+```java
+    getFirstName()
+```
+
+- Get customer `lastName`. The return value will be a `String`.
+```java
+    getLastName()
+```
+
+- Get customer `customerType`. The return value will be a `String`.
+```java
+    getCustomerType()
+```
+
+- Get customer `bikes`. The return value will be a `List<UUID>`.
+```java
+    getBikes()
 ```
