+++ rmi-customer/pom.xml
@@ -26,7 +26,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>web.service.rmi.customer.CustomerService</mainClass>
+                            <mainClass>web.service.rmi.interfaces.customer.CustomerService</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
