+++ rmi-customer/src/main/java/models/customer/CustomerList.java
@@ -1,0 +1,40 @@
+package models.customer;
+
+import models.CustomerType;
+import interfaces.customer.CustomerService;
+import interfaces.customer.CustomerListService;
+
+import java.rmi.RemoteException;
+import java.rmi.server.UnicastRemoteObject;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+public class CustomerList extends UnicastRemoteObject implements CustomerListService {
+    private final Map<UUID, Customer> customerMap = new ConcurrentHashMap<>();
+
+    public CustomerList() throws RemoteException {
+        super();
+    }
+
+    @Override
+    public void add(String firstName, String lastName, CustomerType customerType) throws RemoteException {
+        Objects.requireNonNull(firstName);
+        Objects.requireNonNull(lastName);
+        Objects.requireNonNull(customerType);
+
+        boolean create;
+
+        do {
+            var uuid = UUID.randomUUID();
+            create = customerMap.putIfAbsent(uuid, new Customer(uuid, firstName, lastName, customerType)) == null;
+        } while (!create);
+    }
+
+    @Override
+    public Optional<CustomerService> getCustomerByUUID(String uuid) throws RemoteException {
+        return Optional.ofNullable(customerMap.get(UUID.fromString(Objects.requireNonNull(uuid))));
+    }
+}
